pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-retriever"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "cython>=3.0.12",
    "fastapi>=0.115.12",
    "fugashi[unidic]>=1.4.0",
    "mistune>=3.1.3",
    "protobuf>=6.30.2",
    "pydantic-settings>=2.7.1",
    "pyterrier-dr>=0.6.2",
    "python-terrier>=0.13.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.1.0",
    "sentencepiece>=0.2.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "obret",
]
namespaces = false

[tool.setuptools]
provides = ["obret"]

[tool.ruff.lint]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]
